Postgres needs a fetch size and autocommit off to stream correctly
The db.stream call in ByteArrayJournalDao.messages can ask for all
messages for a certain persistenceId. For example when a Persistent
actor is restoring its state. Against a Postgres database this caused
the entire table to be loaded into a result set in memory before the
first message was delivered to the actor.

The fix for this is to turn off auto-commit (using .transactionally) and
to set a fetchSize. With these settings the Postgres driver delivers
the rows in batches.

Without this fix, and on a large table, the Persistent actor will time
out the recovery phase as no messages are delivered to it.

See https://stackoverflow.com/a/31371423/

# Conflicts:
#	src/main/scala/akka/persistence/jdbc/journal/dao/ByteArrayJournalDao.scala